Multiplayer: Moved client and server code to separate packages
